Add files via upload
This Python-Script will be later used to compress a picture like JPEG does, so this script consists of two functions: the discrete cosine transformation and the inverse discrete cosine transformation. 
They both take a 2-dimensional array with the structure: 
 array[row][column] 
The array contains 8 rows and a row 8 value.
Both functions return an array with a similar structure like their input.