Update and rename complexNumbersTest.py to oneDimensionalDFT.py
Functions for a one-dimensional DFT and iDFT were added. Also, functions that rounded the final values were added. Testing the functions showed that the calculation wasn't precise by 100%. This is not because of a wrong formula or wrong implementation, but because of the number of digits a complex number can store after the floating point.