Add files via upload
This Python script is used to test if the implementation of the 2-D-DFT function is correct. The size of the matrix for the input of the 2-D-DFT-function is 2x2. Both the normal and the inverse function seem to work.
It might be, that you get a little different result if you take a matrix, transform it and then transform it back. This is due to the number of digits a complex number in Python can store behind the floating point.