lossy Image compression added (not complete yet)
A Python script is added which takes a BMP picture, converts it into YCbCr, does chroma subsampling, a discrete Fourier transform. Then it does all the inverse functions and creates a new BMP picture. Be aware of the fact that it only processes BMP pictures, which have three layers (RGB) and which widths and heights are multiples of 8 (if you want to do the 4:2:0 color subsampling they have to be multiples of 16). The "lossyImageCompression.py" is the main function. Also please note, that you need the Python Image Library (or how it is called now) and please check first if it works. Without the Python Image Library the "lossyImageCompression.py" and the "extractingPixelDataOF.py" won't work. The "OF" in the end of the scripts name indicates,  that this file only contains functions and no main program.