Update lossyImageCompression.py
Now the quantization works, the problem was wrong used variable names. Also, the name for the new image was changed, to be able to change the structure inside the project, e.g. by putting the new image into another folder. But you have to use absolute paths because Python causes some problems if you want to use relative paths.